moving compile into its own module
