+++ src/wf_core/creator.rs
@@ -2,6 +2,7 @@ use super::basic_message::BasicMessage;
 use super::segment::MessageSegment;
 use super::types::MessageType;
 use super::wf_buffer::common::{crop_bits, decode_from_hexadecimal, to_hex};
+use crate::wf_convert::FieldValue;
 
 pub const PREFIX: &str = "WF";
 pub const PROTOCOL_VERSION: &str = "1";
@@ -10,22 +11,8 @@ pub const FIELD_VERSION: &str = "Version";
 pub const FIELD_MESSAGETYPE: &str = "MessageCode";
 pub const FIELD_TESTMESSAGETYPE: &str = "PseudoMessageCode";
 
-pub fn compile<T: AsRef<str> + Into<String>>(data: &[T]) -> BasicMessage {
-    let mut header: MessageSegment = MessageSegment::generic_header_segment();
-    header.set_all(data.as_ref(), 0);
-
-    let mut message_type = get_message_type(&header);
-    let body_start_index = header.get_number_of_fields();
-    let body = &mut message_type.body;
-
-    //need switch statement here
-
-    body.set_all(data.as_ref(), body_start_index);
-    BasicMessage::new(message_type.message_code, header, message_type.body)
-}
-
-pub fn encode<T: AsRef<str> + Into<String>>(fields: &[T]) -> String {
-    let basic_message = compile(fields);
+pub fn encode<T: FieldValue>(fields: &[T]) -> String {
+    let basic_message: BasicMessage = fields.into();
     let (message_encoded, len) = basic_message.encode();
 
     to_hex(&crop_bits(message_encoded, len as isize))
